* Translations/es.po: Updated Spanish translations * Translations/messages.po: * Translations/Translations.mdse: * LongoMatch/IO/SectionsWriter.cs: the hotkey values must be writted as int values * LongoMatch/Gui/SectionsPropertiesWidget.cs: Speed improvments adding/deleting sections

+++ LongoMatch/IO/SectionsWriter.cs
@@ -88,8 +88,8 @@ namespace LongoMatch.IO
 				sb.Append(String.Format("<add key=\"Red{0}\" value=\"{1}\" />",i,tn.Color.Red));
 				sb.Append(String.Format("<add key=\"Green{0}\" value=\"{1}\" />",i,tn.Color.Green));
 				sb.Append(String.Format("<add key=\"Blue{0}\" value=\"{1}\" />",i,tn.Color.Blue));
-				sb.Append(String.Format("<add key=\"Modifier{0}\" value=\"{1}\" />",i,tn.HotKey.Modifier));
-				sb.Append(String.Format("<add key=\"Key{0}\" value=\"{1}\" />",i,tn.HotKey.Key));
+				sb.Append(String.Format("<add key=\"Modifier{0}\" value=\"{1}\" />",i,(int)(tn.HotKey.Modifier)));
+				sb.Append(String.Format("<add key=\"Key{0}\" value=\"{1}\" />",i,(int)(tn.HotKey.Key)));
 				i++;
 			}			
 			sb.Append("</configuration>");
